Completely refactor scene management
There's now exactly one offline and one online scene, which are special base scenes that others are loaded additively into.

Additionally, we apparently need to be using SceneMessage from the server to tell the client what to do--this does a bunch of stuff behind the scenes that doesn't happen otherwise.